fix(Structues): rename old module's name
